Fix new warnings on Windows

+++ Baikal/SceneGraph/material.cpp
@@ -218,14 +218,14 @@ namespace Baikal
         return m_inputs.size();
     }
 
-    Material::Input Material::GetInput(std::uint32_t idx) const
+    Material::Input Material::GetInput(std::size_t idx) const
     {
         if (idx >= GetNumInputs())
             throw std::logic_error(
                 "Material::GitInputByIndex(...): idx can not be bigger than number of inputs");
 
         auto iter = m_inputs.begin();
-        for (std::uint32_t i = 0; i < idx; i++)
+        for (std::size_t i = 0; i < idx; i++)
             ++iter;
 
         return iter->second;
